Fix progress indicator CPU usage
This is a bit of a hack, but seems there is not way to set the fps of
ClutterTimeline, so new-frame cb is called very fast. Redrawing the
indicator on every new frame doesn't make much sense, so check if we really
need to redraw it.

Also, fix some remaining compiler warnings while at it.

+++ src/launcher/hd-launcher-page.c
@@ -579,7 +579,7 @@ hd_launcher_page_update_emptylabel (HdLauncherPage *page, gboolean is_portrait)
 
   HdLauncherPagePrivate *priv = HD_LAUNCHER_PAGE_GET_PRIVATE (page);
   guint x1, y1;
-  guint label_width, label_height;
+  gfloat label_width, label_height;
   guint page_width = is_portrait ? HD_LAUNCHER_PAGE_HEIGHT : HD_LAUNCHER_PAGE_WIDTH;
   guint page_height = is_portrait ? HD_LAUNCHER_PAGE_WIDTH : HD_LAUNCHER_PAGE_HEIGHT;
